+++ CHANGELOG.md
@@ -651,6 +651,7 @@
 - [Added `Decimal.round`.][9672]
 - [Implemented write support for Enso Cloud files.][9686]
 - [Added `Integer.to_decimal` and `Float.to_decimal`.][9716]
+- [Added `Decimal.floor`, `.ceil`, and `.trunc`.][9694]
 - [Added `recursive` option to `File.delete`.][9719]
 
 [debug-shortcuts]:
@@ -952,6 +953,7 @@
 [9663]: https://github.com/enso-org/enso/pull/9663
 [9672]: https://github.com/enso-org/enso/pull/9672
 [9686]: https://github.com/enso-org/enso/pull/9686
+[9694]: https://github.com/enso-org/enso/pull/9694
 [9716]: https://github.com/enso-org/enso/pull/9716
 [9719]: https://github.com/enso-org/enso/pull/9719